% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop}
\alias{pop}
\title{County-Level Population, 1969-2016}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 149766 rows and 5 columns.}
\source{
\url{https://seer.cancer.gov/popdata/download.html}
}
\usage{
pop
}
\description{
County-level population from SEER
}
\keyword{datasets}
