% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_since_1990}
\alias{pop_since_1990}
\title{County-Level Population, 1990-2016}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 84718 rows and 16 columns.}
\source{
\url{https://seer.cancer.gov/popdata/download.html}
}
\usage{
pop_since_1990
}
\description{
County-level population from SEER
}
\keyword{datasets}
