% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_since_1990_adjusted}
\alias{pop_since_1990_adjusted}
\title{County-Level Population, 1990-2016, Adjusted}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 84719 rows and 16 columns.}
\source{
\url{https://seer.cancer.gov/popdata/download.html}
}
\usage{
pop_since_1990_adjusted
}
\description{
County-level population from SEER, adjusted for the population shifts due to hurricanes Katrina (August 29) and Rita (September 24).
}
\keyword{datasets}
